backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[argon2]
mediapipe
opencv-python-headless
numpy
pillow
argon2-cffi
